// utils/api/productSessions.ts
export interface PublicProductSession {
  productId: string; // MongoDB ObjectId string
  sessionId: number;
  count: number;
  price: string;
  finalPrice: string;
  discountPrice: string;
  discount: number;
  perSession: number;
}